flask-meter-ocr/requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-Migrate>=4.0.0
Flask-JWT-Extended>=4.5.0
PyMySQL>=1.1.0
Pillow>=10.0.0
pytesseract>=0.3.10
Werkzeug>=2.3.0
python-dotenv>=1.0.0
